refactor(node): move RadManager logic outside of handlers
